refactor(#477 PR 3c-ii-a): move auto-config helpers to src/server/integrations/apply.ts
Library factor — first sub-PR of #477 PR 3c-ii. Pure code move, zero
behavior change. Pre-requisite for the schema bump and wizard apply
path in 3c-ii-b.

What moved (`src/cli/setup.ts` → `src/server/integrations/apply.ts`):
- applyConfig, applyConfigWithToken
- detectTargets (including MSIX detection)
- buildMcpEntries
- validateChannelShimPrereq
- installSkill
- atomicWrite helper
- TargetKind, DetectedTarget, McpEntry, McpEntries, BuildMcpEntriesOptions, DetectOptions types
- PACKAGE_ROOT, CHANNEL_DIST constants (re-exported for runSetup)

What stayed in `src/cli/setup.ts`:
- runSetup (CLI orchestration; rewritten in 3c-ii-c as non-interactive `--apply` wrapper)
- Back-compat re-exports so `tandem rotate-token`, existing tests, and
  any other consumer keep importing from `src/cli/setup.js` unchanged.
  The re-export layer is removed in 3c-ii-c.

Direct-import callers updated:
- src/server/integrations/existing-config.ts → ./apply.js
- src/server/mcp/routes/setup.ts → ../../integrations/apply.js

Plan: docs/plans/477-pr-3c-ii-auto-config-removal.md (revised after three
adversarial reviews — security, contrarian, completeness — that surfaced
path-traversal mitigation, npm CLI first-run gap, schema bump treatment,
and ~17 doc surfaces still to update in 3c-ii-c).

Verification: `npm run typecheck` clean. Full vitest suite (187 files,
2344 passed / 11 skipped) green.
